requirements.txt
pycocotools
torch
torchvision
pytorch-ignite
matplotlib
pyyaml
nltk
click